+++ RELEASE_NOTES.md
@@ -18,6 +18,8 @@ Release Notes
     * [SER-1924] - Allow some transaction types clickable
     * [SER-1524] - Fix pagination issue
     * [SER-1916] - Create a UI for repayments download
+    * [SER-1112] - Create UI for field configuration (CRUD)
+
 
 ## Version 1.0.0 - for use with Fineract Web App